changes to fetcher

+++ trader/src/indicators/macd_histogram.rs
@@ -1,13 +1,15 @@
 use super::{Indicator, EMA, MACD};
 use crate::economy::Monetary;
 
-pub struct MACDHistogram<const SHORT: u64, const LONG: u64, const PERIOD: u64> {
+pub struct MACDHistogram<const SHORT: usize, const LONG: usize, const PERIOD: usize> {
     histogram: Monetary,
     macd: MACD<SHORT, LONG>,
     signal: EMA<PERIOD>,
 }
 
-impl<const SHORT: u64, const LONG: u64, const PERIOD: u64> Indicator for MACDHistogram<SHORT, LONG, PERIOD> {
+impl<const SHORT: usize, const LONG: usize, const PERIOD: usize> Indicator
+    for MACDHistogram<SHORT, LONG, PERIOD>
+{
     type Output = Option<(Monetary, Monetary)>;
 
     fn initialize(value: Monetary) -> Self {
@@ -18,7 +20,7 @@ impl<const SHORT: u64, const LONG: u64, const PERIOD: u64> Indicator for MACDHis
         }
     }
 
-    fn evaluate(&mut self, value: Monetary) -> Self::Output {        
+    fn evaluate(&mut self, value: Monetary) -> Self::Output {
         if let Some(macd) = self.macd.evaluate(value) {
             if let Some(signal) = self.signal.evaluate(macd) {
                 self.histogram = macd - signal;
